[TD-497] FileContentTypeUtils 메소드들을 FileContentType으로 이동
