+++ threedollar-api-core/src/main/java/com/depromeet/threedollar/application/provider/upload/UploadProvider.java
@@ -1,7 +1,6 @@
 package com.depromeet.threedollar.application.provider.upload;
 
 import com.depromeet.threedollar.application.provider.upload.dto.request.UploadFileRequest;
-import com.depromeet.threedollar.common.utils.FileContentTypeUtils;
 import com.depromeet.threedollar.external.client.filestorage.FileStorageClient;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
@@ -14,7 +13,7 @@ public class UploadProvider {
     private final FileStorageClient fileStorageClient;
 
     public String uploadFile(UploadFileRequest request, MultipartFile file) {
-        FileContentTypeUtils.validateAvailableContentType(file.getContentType(), request.getType());
+        request.validateAvailableContentType(file.getContentType());
         String fileName = request.getFileNameWithBucketDirectory(file.getOriginalFilename());
         fileStorageClient.uploadFile(file, fileName);
         return fileStorageClient.getFileUrl(fileName);
