+++ threedollar-api-core/src/main/java/com/depromeet/threedollar/application/provider/upload/dto/request/UploadFileRequest.java
@@ -6,6 +6,10 @@ public interface UploadFileRequest {
 
     FileType getType();
 
+    default void validateAvailableContentType(String contentType) {
+        getType().validateAvailableContentType(contentType);
+    }
+
     default String getFileNameWithBucketDirectory(String originalFileName) {
         return getType().createUniqueFileNameWithExtension(originalFileName);
     }
